+++ _schemaV2.graphql
@@ -3172,6 +3172,44 @@ type BidIncrementsFormatted {
 # exceed the size of a 32-bit integer, it's encoded as a string.
 scalar BigInt
 
+type BulkUpdatePartnerArtworksMutationFailure {
+  mutationError: GravityMutationError
+}
+
+input BulkUpdatePartnerArtworksMutationInput {
+  # Whether Artsy domestic shipping should be enabled
+  artsyShippingDomestic: Boolean
+
+  # Whether Artsy international shipping should be enabled
+  artsyShippingInternational: Boolean
+  clientMutationId: String
+
+  # ID of the partner
+  id: String!
+
+  # The partner location ID to assign
+  location: String
+}
+
+type BulkUpdatePartnerArtworksMutationPayload {
+  bulkUpdatePartnerArtworksOrError: BulkUpdatePartnerArtworksMutationType
+  clientMutationId: String
+}
+
+type BulkUpdatePartnerArtworksMutationSuccess {
+  skippedPartnerArtworks: BulkUpdatePartnerArtworksResponse
+  updatedPartnerArtworks: BulkUpdatePartnerArtworksResponse
+}
+
+union BulkUpdatePartnerArtworksMutationType =
+    BulkUpdatePartnerArtworksMutationFailure
+  | BulkUpdatePartnerArtworksMutationSuccess
+
+type BulkUpdatePartnerArtworksResponse {
+  count: Int
+  ids: [String]
+}
+
 type BuyersPremium {
   # A formatted price with various currency formatting options.
   amount(
@@ -10561,6 +10599,11 @@ type Mutation {
   adminUpdateFeatureFlag(
     input: AdminUpdateFeatureFlagInput!
   ): AdminUpdateFeatureFlagPayload
+
+  # Update all artworks that belong to the partner
+  bulkUpdatePartnerArtworks(
+    input: BulkUpdatePartnerArtworksMutationInput!
+  ): BulkUpdatePartnerArtworksMutationPayload
   captureHold(input: CaptureHoldInput!): CaptureHoldPayload
   commerceAddInitialOfferToOrder(
     # Parameters for AddInitialOfferToOrder
@@ -10905,11 +10948,6 @@ type Mutation {
   updateNotificationPreferences(
     input: updateNotificationPreferencesMutationInput!
   ): updateNotificationPreferencesMutationPayload
-
-  # Update all artworks that belong to the partner
-  updatePartnerArtworks(
-    input: UpdatePartnerArtworksMutationInput!
-  ): UpdatePartnerArtworksMutationPayload
   updateSavedSearch(input: UpdateSavedSearchInput!): UpdateSavedSearchPayload
   updateSmsSecondFactor(
     input: UpdateSmsSecondFactorInput!
@@ -11715,11 +11753,6 @@ type PartnerArtworkGrid implements ArtworkContextGrid {
   ctaHref: String
   ctaTitle: String
   title: String
-}
-
-type PartnerArtworksBulkEditErrors {
-  count: Int
-  ids: [String]
 }
 
 type PartnerCategory {
@@ -15482,43 +15515,6 @@ type updateNotificationPreferencesMutationPayload {
     authenticationToken: String
   ): [NotificationPreference!]!
 }
-
-type UpdatePartnerArtworksMutationFailure {
-  mutationError: GravityMutationError
-}
-
-input UpdatePartnerArtworksMutationInput {
-  # Whether Artsy domestic shipping should be enabled
-  artsyShippingDomestic: Boolean
-
-  # Whether Artsy international shipping should be enabled
-  artsyShippingInternational: Boolean
-  clientMutationId: String
-
-  # ID of the partner
-  id: String!
-
-  # The partner location ID to assign
-  location: String
-}
-
-type UpdatePartnerArtworksMutationPayload {
-  clientMutationId: String
-  partnerArtworksBulkEditOrError: UpdatePartnerArtworksMutationType
-}
-
-type UpdatePartnerArtworksMutationSuccess {
-  partnerArtworksBulkEdit: UpdatePartnerArtworksMutationSuccessDetails
-}
-
-type UpdatePartnerArtworksMutationSuccessDetails {
-  errors: PartnerArtworksBulkEditErrors
-  success: Int
-}
-
-union UpdatePartnerArtworksMutationType =
-    UpdatePartnerArtworksMutationFailure
-  | UpdatePartnerArtworksMutationSuccess
 
 # Autogenerated input type of UpdateSavedSearch
 input UpdateSavedSearchInput {
