+++ src/schema/v2/bulkUpdatePartnerArtworksMutation.ts
